Add support for controlling individual sink inputs
There is much to critic about the implementation and I'm aware of
that.

For one, it uses a global variable, so it can know which index to
fiddle with, but it is worse than that, because if the
index (*pavol-index*) is zero, it assumes it's a sink, otherwise, it
assumes it's an input sink.